[FIX] user unread message 내부 로직 변경

+++ backend/src/main/java/com/musicat/data/dto/user/UserUnreadMessageDto.java
@@ -9,5 +9,5 @@ import lombok.*;
 @NoArgsConstructor
 @Builder
 public class UserUnreadMessageDto {
-    private int userUnreadMessage;
+    private long userUnreadMessage;
 }
